Move RPC server logic into separate project
