+++ app/oracle-server/src/main/scala/org/bitcoins/oracle/server/OracleServerMain.scala
@@ -1,7 +1,7 @@
 package org.bitcoins.oracle.server
 
 import org.bitcoins.dlc.oracle.config.DLCOracleAppConfig
-import org.bitcoins.server.{BitcoinSRunner, Server}
+import org.bitcoins.server.routes.{BitcoinSRunner, Server}
 
 import scala.concurrent.Future
