+++ app/server/src/main/scala/org/bitcoins/server/BitcoinSServerMain.scala
@@ -3,6 +3,7 @@ package org.bitcoins.server
 import akka.actor.ActorSystem
 import akka.dispatch.Dispatchers
 import akka.http.scaladsl.Http
+import grizzled.slf4j.Logging
 import org.bitcoins.chain.blockchain.ChainHandler
 import org.bitcoins.chain.config.ChainAppConfig
 import org.bitcoins.chain.models._
@@ -18,13 +19,15 @@ import org.bitcoins.feeprovider._
 import org.bitcoins.node._
 import org.bitcoins.node.config.NodeAppConfig
 import org.bitcoins.node.models.Peer
+import org.bitcoins.server.routes.{BitcoinSRunner, Server}
 import org.bitcoins.wallet.Wallet
 import org.bitcoins.wallet.config.WalletAppConfig
 
 import scala.concurrent.{ExecutionContext, Future, Promise}
 
 class BitcoinSServerMain(override val args: Array[String])
-    extends BitcoinSRunner {
+    extends BitcoinSRunner
+    with Logging {
 
   override val actorSystemName = "bitcoin-s-server"
 
@@ -288,26 +291,24 @@ class BitcoinSServerMain(override val args: Array[String])
     val server = {
       rpcPortOpt match {
         case Some(rpcport) =>
-          Server.apply(conf = nodeConf,
-                       handlers =
-                         Seq(walletRoutes, nodeRoutes, chainRoutes, coreRoutes),
-                       rpcbindOpt = rpcbindOpt,
-                       rpcport = rpcport)
+          Server(conf = nodeConf,
+                 handlers =
+                   Seq(walletRoutes, nodeRoutes, chainRoutes, coreRoutes),
+                 rpcbindOpt = rpcbindOpt,
+                 rpcport = rpcport)
         case None =>
           conf.rpcPortOpt match {
             case Some(rpcport) =>
-              Server.apply(
-                conf = nodeConf,
-                handlers =
-                  Seq(walletRoutes, nodeRoutes, chainRoutes, coreRoutes),
-                rpcbindOpt = rpcbindOpt,
-                rpcport = rpcport)
+              Server(conf = nodeConf,
+                     handlers =
+                       Seq(walletRoutes, nodeRoutes, chainRoutes, coreRoutes),
+                     rpcbindOpt = rpcbindOpt,
+                     rpcport = rpcport)
             case None =>
-              Server.apply(
-                conf = nodeConf,
-                handlers =
-                  Seq(walletRoutes, nodeRoutes, chainRoutes, coreRoutes),
-                rpcbindOpt = rpcbindOpt)
+              Server(conf = nodeConf,
+                     handlers =
+                       Seq(walletRoutes, nodeRoutes, chainRoutes, coreRoutes),
+                     rpcbindOpt = rpcbindOpt)
           }
       }
     }
